cli: fix edge case in pure punctuation aliases
Summary:
The command tokenizer was treating a string of punctuation characters at the start of a command as a single unit without the possibility of subdividing it. This meant that, for example, the PHP command

=$x

would only evaluate an alias if the alias was =$, and ignore the '=' for print alias.

Change the code to find the maximal punctuation alias instead.

Reviewed By: ebluestein

Differential Revision: D8934078

fbshipit-source-id: 3f5075132b9be720697d88992ecbff946904a949