Catch and log errors in BigDigServer
Summary:
There is a scary comment in `BigDigServer` about not handling WS/HTTPS server errors, and I think I managed to encounter one. I'm not sure how to reproduce it, but we can at least see if this is happening in the wild.

I thought this might've been related to https://github.com/websockets/ws/issues/1256 but we're still using ws 3.2.0 so that wasn't exactly it. But I could imagine the HTTPS server erroring in other cases maybe?

I assigned a logger category so that these are easier to find.

Reviewed By: wanderley

Differential Revision: D8954324

fbshipit-source-id: 9582ebb9a394eafbb52aba1584998ef6c4523ca2