Enhance documentation for commons-node/passesGK
Summary: Many Nuclide packages do not run on the same machine or user as the Nuclide/Atom client. It may not be obvious that a `passesGK` call will use `os.userInfo` from the Node.js standard API and thus may have unexpected results if called on a language server running in a different environment than the user.

Reviewed By: wbinnssmith

Differential Revision: D9004512

fbshipit-source-id: 70192f3977e11918cd937dc549d752f68bc74cba