Rename onConnectionEnd to onClientClose to Better Describe its Purpose
Summary: The old function name `onConnectionEnd` sounds like including `onUnexpectedConnectionEnd` or in other word, it causes confusion that `onUnexpectedConnectionEnd` is just a special case of `onConnectionEnd`. To avoid confusion and better describe what this function is supposed to do, in this diff, I renamed it to `onClientEnd` which is apparently different from `onUnexpectedConnectionEnd`.

Reviewed By: wanderley

Differential Revision: D8798665

fbshipit-source-id: 9b026cb773854140862c066b5545c92a5b448c2f