don't re-emit errors from the Proxy
Summary:
Not sure why I was doing this to begin with, but it's definitely wrong.

I'm not sure this will fix @[1490863821:logand]'s issue since it appears as though Metro is refusing the connections. We'll have to dig in a little more to better understand why Metro stops allowing connections.

Reviewed By: hansonw

Differential Revision: D9004389

fbshipit-source-id: 09e0278bfbf6269092da90a0e5e3858c2388c1d2