Include working sets as a first-class projects feature
Summary:
We're currently iterating on the behavior of projects and their relationship to working sets.

Initially, we simply mounted the `paths` in Atom, however, this isn't ideal for many use cases at Facebook—particularly www—which rely on extra-project paths being browsable.

The current approach (prior to this diff) is to use the project's `paths` to determine the repository root. That directory was then mounted and then a special working set was created with the `paths`. However, this precludes a workflow like Nuclide's where we want to completely ignore files outside of our `paths`.

This diff switches to a different model. Like with the original approach, `paths` are mounted. Working sets are now a first-class projects feature and can be defined in the project file:

```
paths = ['path/to/nuclide']

[[workingSets]]
name = 'File Tree'
paths = ['path/to/nuclide/pkg/nuclide-file-tree']
initiallyActive = true

[[workingSets]]
name = 'Packages'
paths = ['path/to/nuclide/pkg']
```

Like with the second approach, these working sets are (currently, at least) uneditable and displayed in a separate section in the working set modal.

Pros:
1. Model is much simpler and doesn't require changing the default behavior of `paths`.
2. Users can share any number of working sets.
3. More expressivity means supporting more workflows.

Cons:
1. Because there can be n working sets, we no longer have a single "primary" set of paths. This might be important to other tools, but that's only hypothetical.
2. Projects are conceptually more complex.
3. We now have to worry about supporting multi-root projects again.

Possibly Pro, Possibly Con:
1. Working sets UX needs work. This would drive that, which could be a good thing.

Reviewed By: hansonw

Differential Revision: D8667367

fbshipit-source-id: 4ac134bf40a57c913adf415ac77e3e3bd3d4c95b