Start stabilizing the order of File Tree context menu items
Summary:
Back in February I [added support](https://github.com/atom/atom/pull/16661) in Atom for controling the order of context menu items. The API is not particularly pretty, but it works.

Responding to alobbFB's report in T26643515, I've added our first use of this API.

You'll notice that in this diff, `tree-view:search-in-directory` does not declare its own position. Instead, other entries position themselves relative to it. This is because in some cases (I'm not sure if this is intentional or not) `atom/find-and-replace` provides a context menu entry by the same name.

To handle this, "Show in Finder" and "Set to Current Working Root" define themselves relative to both `project-find:show-in-current-directory` and `tree-view:search-in-directory`.

This way, whichever entry wins, the ordering is stable.

{F134321737}

Reviewed By: matthewwithanm

Differential Revision: D9070332

fbshipit-source-id: 76dd805ae739e78b8e8275475f1b9e3356b83817