Restore focus, *then* perform action side effects
Summary:
Rather than the other way around. This fixes situations where toggle commands hide dock contents, but then focus is restored to them and the dock opens immediately again.

This also removes a check in `showModal` which asserted that focus hadn't been moved somewhere else, as this was necessary to handle atom commands which shifted focus elsewhere. Now that effects are applied *after* the modal closes, we no longer need this check.

Reviewed By: matthewwithanm

Differential Revision: D8980183

fbshipit-source-id: 6103172e03f0f8c1c273edb168a9657dbc30df89