Return Observable for LSP rename
Summary:
On especially large refactorings (i.e. on `nullthrows` in `www`), rename will timeout after 60 seconds (as set by RPC).

However, `find-references` does not timeout on especially large requests for references. This is due to the fact that find-references returns an `Observable`.

As a result, rename LSP has been modified to return an `Observable` rather than a `Promise`.

Reviewed By: arxanas

Differential Revision: D9112875

fbshipit-source-id: df73b1586236159a41ba1499a313776c591ecff7