Fix GK check in nuclide-js-imports-server
Summary: `nuclide-js-imports-server` was responsible for checking the GK related to component autocompletion and documentation on hover. This means that it could have a different Unix name than the one running the Nuclide/Atom client. This commit moves that logic to the `nuclide-js-imports-client` package and adds it to the LSP initialization options. This will be moved to configuration pulling as per the LSP spec in the near future (https://microsoft.github.io/language-server-protocol/specification#workspace_configuration).

Reviewed By: hansonw

Differential Revision: D9029049

fbshipit-source-id: 9a6c188927ec74856593eae2db98f3a7d6083bac