don't crash server when there's no service listening for a reverse tunnel
Summary:
Calling `socket.destroy` with an error actually causes the error to be emitted, but it is emitted on the next tick, after we remove the listeners. :(

In this diff, I stopped calling `destroy` with the error since we've already handled it at the point we call `destroy`.

One side effect, is that the `close` event gets sent to the server after the error, but we've already closed the socket. So I print a warning in that case.

Reviewed By: stepanhruda

Differential Revision: D9125708

fbshipit-source-id: 81463515e771486fc42113f6e2c944b0e4b3d541