Speed up hg log by reducing fields requested
Summary:
`hg log` in vcs log was running `hg log -Tjson` which pulls in several unnecessary fields.

a typical hg log call in ovrsource for 100 commits (the default) is ~100K. This change brings this down 20% to 80K by not transferring extra fields over the rpc layer. This makes it faster, and means we could be more aggressive about how many results we show.

I thought this was the issue causing maxBuffer exceeded, but that's actually in the diff part, which I'll fix in other diffs

Reviewed By: tjfryan

Differential Revision: D9156174

fbshipit-source-id: 8ccd5f8f480ad28417275d6e7b46f41e0f220346