Show error when loading vcs log diff inline
Summary: If loading a vcs log diff gives an error, we don't need to crash the whole component. We can just show the error instead of the diff viewers. Then maybe you can try again on a different diff. (the most common error is that there's too many changes to show)

Reviewed By: tjfryan

Differential Revision: D9162504

fbshipit-source-id: 80c3f6f86783e4299bf1b06e3f67359e8138f822