Fix state issues on restart
Summary: In launch mode, when the program exits, we set up the session again. We do the same thing when the adapter exits, in case it crashes. We need to check to make sure both of these things don't fire at once, because restarting the sesion kills the old adapter.

Reviewed By: ebluestein

Differential Revision: D9180947

fbshipit-source-id: ccfcefa4f340f0822ba7ed3ce2f8ca094892be89