Add FileSystemService/readdirRecursive - returns a flattened list of recursively found directory entries
Summary:
This adds a `readdirRecursive` function which recursively traverses a file tree (with the provided limit) and returns a flattened list of directory entries, relative to the start path. The entries are sorted to ensure that the limit is applied consistently.

One potential pitfall with the current implementation is an inability to deal with circular symlinks.

Reviewed By: hansonw

Differential Revision: D9113862

fbshipit-source-id: ac4717cacb8f9f5ff4bd7c6824cecead62b3c73c