File Tree: Fix disappearing nodes when revealing in file tree
Summary:
Okay, so this is a bit of a weird one. The symptom was that "Reveal in File Tree" would make nodes disappear.

The reason was that, due to an oversight allowed by frustrating type unsafety, the `FileTreeNode` constructor was sometimes passed a `Store` instead of a `StoreConfigData` so all of the conf property lookups, including `shouldBeShown` would return `undefined`.

This should have been caught by Flow but, for some reason, it was treating the `Store` type in this module (whose import was accidentally omitted I guess) as `any`.

Reviewed By: aadisriram

Differential Revision: D9192269

fbshipit-source-id: 45c244e9a78bd0dd88c49b01fd2ffe2fe7632afd