fix edge case on 'enter'
Summary: If a user hit 'enter' on a command, and the command wrote text to stdout during the callback, the cursor would not be in the proper position on the next prompt (it would end up where it was had the command text still been there). Fix it to be at the end of the prompt.

Reviewed By: ebluestein

Differential Revision: D9198642

fbshipit-source-id: 3c3d33092eaf3748046232e9732fcf8668d21151