Copy filename in commit info view without spaces
Summary:
If you right click to copy a filename in the commit info view, it would also copy the zero-width spaces used to break at `/`. This is annoying if you're actually trying to use the filename you copied. So we can add a custom copy handler that filters the spaces and then writes it to the clipboard.

Note: I tried and failed to get this to work with `core:copy`. Since you can't really select and use cmd-c, there's not a big benefit to using `core:copy` anyway. But if you know how to force the custom copy behavior, let me know!

Reviewed By: aadisriram

Differential Revision: D9199740

fbshipit-source-id: db3c7126e3be114c127ff5e9dd1aa99736fd90b7