File tree: Fix Reveal in File Tree
Summary:
"Reveal in File Tree" had some issues. The first time it would work, but subsequent times it would fail to scroll to revealed files. Even before the Redux refactor, it would need to be triggered twice to work.

It turns out there was a lot of complicated logic in here for trying to determine whether scrolls were user-initiated which I think was unnecessary. Basically, a big loop whereby the scroll events of the `List` were observed and then fed back to it via `FileTreeSidebarComponent`. There doesn't seem to be any downside to just waiting until we've rendered the node to clear it.

Note that, although `List` uses a declarative API for setting the `scrollToIndex`, it's really more of an imperative side-effect. If it's set and not changed, for example, it doesn't lock scrolling or anything—only the change seems to have any effect.

Reviewed By: captbaritone

Differential Revision: D9211490

fbshipit-source-id: 3ff7fa6375248139a33d061efb2f0b98b5948746