fix leftover input when exiting
Summary: On exit, the debugger would leave a partial escape sequence at the shell prompt. This was because of a cursor position request which we didn't wait for the answer to, because of a message being written to the console while exiting. Removing the message (which didn't print anyway) fixes the problem.

Reviewed By: ebluestein

Differential Revision: D9232161

fbshipit-source-id: b0afd3fe8884499d386b94899694e68eeab6a908