Create ConfigManager Class
Summary: In order to use an atom$Config for non-user-facing values, a separate config from the global atom.config is needed. This diff pulls out the feature-config functions into a reusable class, wrapping a provided atom$Config, and changes feature-config to use it. In a later diff, the same class will be used to implement nuclide-config

Reviewed By: wbinnssmith

Differential Revision: D9069382

fbshipit-source-id: 30e65245138f95c9f8eddabbf1b166a03b55cdb3