add SimpleConnect API for connections that don't need filesystem access
Summary:
The Miso team needs a simple way to get a `BigDigClient` (to create a tunnel) without going through the current Nuclide APIs. They won't be creating a filesystem connection, so they don't need a `RemoteConnection`. I created this API with their input.

Right now, we're handing them back a `ServerConnection` object in the `onConnect` delegate method, which they are expected to call `close` on. This is problematic if the user has another connection to the same server or creates one subsequently. If they close that second connection, it will kill the current `ServerConnection` (since it's set up to shut down the `ServerConnection` when there are no more `RemoteConnections`).

This is obviously wrong and needs to be fixed, but I haven't come up with a satisfactory approach that doesn't introduce more technical debt. The right thing to do here is to refactor `ServerConnection` so it's no longer coupled with `RemoteConnection`, and so there is a clearer delineation of responsibility.

Since Miso won't ship for a few more months, we have time to do the refactoring and get this right. In the meantime we can unblock them via this API with the caveat that things probably won't work as expected if they create other connections to the same dev server.

Reviewed By: matthewwithanm

Differential Revision: D9220384

fbshipit-source-id: 9bd4727b7da954f4dfc0feb01e19f627c7d5a2e7