Make stack trace commands more like hphpd
Summary:
Change 'backtrace' to 'where', add the 'frame', 'up' and 'down' commands explicitly. This both makes the commands more like hphpd and makes the abbreviation for the 'breakpoint' command just 'b' instead of 'br'.

Also, update the help command to show the shortest alias that will work for each command, as hphpd does -- this code is shared, so this will work with subcommands as well.

Reviewed By: ebluestein

Differential Revision: D9242796

fbshipit-source-id: 7d50dd83030b41f651acd123ee15b61b68e8d55f