Allow appending to console entries
Summary: This fixes two issues the debugger is having in the console: the first is output events are buffered and not flushed until a \n is seen, which can cause some debugger messages not to appear as expected. The second is single output events that contain embedded \n's (like stack traces) appear as multiple entries in the console, which is messy and unexpected.

Reviewed By: matthewwithanm

Differential Revision: D9244898

fbshipit-source-id: 0066064958cd05a3c762e49c9dc64421ba387e87