Show a dismissable tooltip to inform users that we moved new terminals
Summary:
This displays a dismissable tooltip, which also dismisses when the user closes the terminal or automatically within one minute.

The tooltip informs the user that we have moved the default location for terminals, and that if they move terminals to their preferred location, new ones will open there too.

Reviewed By: matthewwithanm

Differential Revision: D9158776

fbshipit-source-id: 5ee639c5ef258f56627c13c79962c650cfcc401d