Support breakpoint id on stop events in VSP
Summary: Add a support flag to say that the adapter supports returning the breakpoint id (so we can gate features that won't work without it) and returning a breakpoint id with a stop event if possible. Add support to gdb, and showing which breakpoint we're stopped at, in gdb as proof of concept.

Reviewed By: Goom11

Differential Revision: D9271829

fbshipit-source-id: 5d250c792f9f7715a194baf2f7273bba64317692