Fix file becoming stale after hg updating
Summary: When watching files, if the file does not already exist at watch time, we throw an error (due to fstat). This means that the caller ends up with a dead observable that doesn't emit when the file changes (after it gets created). Instead, we can save the list of watched directories, then when we try to watch a file, either fstat or just use the existing path if the file doesn't yet exist.

Reviewed By: wez

Differential Revision: D9322491

fbshipit-source-id: b06979322b8c29f04df0601debcdecd24dae57fa