remove nuclide_prebuilt_keytar gk, which removes the passesGK dependency from keytarWrapper
Summary:
I'm in the process of refactoring fb-interngraph as a module. In the process, I noticed this GK is not needed anymore, and removing it allows me to remove the passesGK dependency.

Talked to @[505887506:hansonw] about it and it looks like the few folks who are holdouts had some issues running chef in the past. Presumably they've resolved those issues by now.

Reviewed By: hansonw

Differential Revision: D9338086

fbshipit-source-id: 8ffc9dff3b376ce79fea956f42d81fcb6e66dd85