Run dev-server-packaging in a forked process
Summary:
This is scratching an itch of mine - it's harder for me to do iterative development because vscode startup is being slowed down.

In development mode, the first thing the extension does is to kick off a task to build an incremental dev package - this drastically lowers the time to apply dev changes to the server. But although it is async, it's CPU intensive and slows down the extension host quite noticeably until the package is built. By running this task in a separate process, I've considerably sped up the startup time.

Reviewed By: bolinfest

Differential Revision: D8042975

fbshipit-source-id: cd553f936675fb658b1081897de45e13c73187af