clean up a couple of confusing messages from gdb
Summary: Internally, the gdb VSP sends a SIGINT to gdb to get it to pay attention to commands. This causes gdb to print messages about the SIGINT and the target stopping, which are confusing to the user because they didn't initiate the stop directly. So filter them out. Also, filter out the hint about 'fbload' because the user won't know that doesn't work at the fbdbg prompt.

Reviewed By: Goom11

Differential Revision: D9321622

fbshipit-source-id: e48d4594ddce239d144bb58be0479e173880db66