Fix: bad behavior on exception in command
Summary: In the case of commands which continue execution, VsDebugSession will send up a "continued" event even if the command eventually fails in the VSP. The "continued" event turns off console input while the target is running, and it's never re-enabled. Practically, this means that if the command fails, the user will never see the debugger prompt again. Catch this case in all the stepping commands.

Reviewed By: Goom11

Differential Revision: D9365460

fbshipit-source-id: ad46f67987c4df4ade7b196c9822319c722993c4