modularize fb-interngraph
Summary:
Move fb-interngraph stuff from commons-node to a module. This unblocks work on modularizing on-demand-core.

We can make this nicer by creating an index and exporting functions so that we don't have to use the full path in every import, but that should be pretty easy now. Maybe a level 1 bootcamp task?

Reviewed By: hansonw

Differential Revision: D9355806

fbshipit-source-id: 87502bb42701809856c9915d423beccb89f2154b