Remote Connection Dialog: Extract ConnectFlow class
Summary:
Switch from a giant function with a lot of closed-over variables to a class.

I have real mixed feelings about this kind of single-method class—especially one that's responsible for its own lifecycle like this. But I think in this case there's no clear way to break out smaller functions and the result is clearer.

Reviewed By: semmypurewal

Differential Revision: D9372973

fbshipit-source-id: 8d56d24122f939d6fdddd0ffd1f3acf8357634cc