disallow cursor position request at exit
Summary: Sending a cursor position request (CPR) causes the TTY driver to write a response to stdin describing where the cursor is. If we send a CPR right as we're about to exit (which happens quite often due to how the debugger command prompt works) we get the response at the shell prompt, which is ugly. Disable CPR's if we know we're exiting.

Reviewed By: ebluestein

Differential Revision: D9412283

fbshipit-source-id: 54cd02b543fe3f5d4d13241bdced2ae00c3e7aee