Expose the FileSystemProvider for big-dig:// to other VS Code extensions.
Summary:
This makes it so that the `FileSystemProvider` for the `big-dig` scheme can be
used by other extensions (providing a workaround for https://github.com/Microsoft/vscode/issues/48034).

Here is an example of how to leverage this:

```
function getFileSystemProvider(): vscode.FileSystemProvider | null {
  const bigDig = vscode.extensions.getExtension("Facebook.big-dig-vscode");
  if (bigDig == null) {
    return null;
  }

  if (!bigDig.isActive) {
    // Alternatively, make this method async and do:
    // return (await bigDig.activate()).fileSystemProvider;
    return null;
  }

  return bigDig.exports.fileSystemProvider;
}

```

Note that this technique of making objects from the `big-dig-vscode` extension could be used more generally
as the basis for a Big Dig client library so that other extensions can talk to Big Dig, as
appropriate. One common case that has been discussed is language extensions reading
from the user's `settings.json` to determine what sorts of extra arguments to add to
the `initializationOptions`.

Reviewed By: siegebell

Differential Revision: D9419766

fbshipit-source-id: d705cc03f71b01e8cefead99a197f8b307c2ad5d