Catch error if creating proxy fails
Summary:
`proxy.startListening()` will get rejected if the listening socket (`Proxy`) encounters an error. This rejected promise is currently unhandled and the entire stack for that connection throws.

We already send `proxyError` to the other side, so we are fine ignoring this error. For verbosity, I added an extra log statement to the server side too - although the client logs its received `proxyError` message as well.

Disclosure: this is currently handling `EADDRINUSE`, which happens before the `listen` callback. It is possible for the server socket to emit an `error` event after it has started listening and the promise has already been successfully resolved. This is a deeper issue with the implementation which I'm not addressing here.

Reviewed By: semmypurewal

Differential Revision: D9559227

fbshipit-source-id: 5195a9fc544f24a30c9ec83ee43197b80396c9c4