remove whitelist
Summary:
I was working with someone on tunnels today and the whitelisting led to some confusion.

Since we've removed the ability to create tunnels manually (unless you're behind a GK) all tunnel creation will go through a code review and therefore there is no longer a need for the whitelist.

Also, I'll be finalizing our move to BigDig in September. Among other things, that means we'll be removing the nuclide-socket-rpc package and refactoring Epics.js so that it only uses BigDig. Let me know if there are any major outstanding issues relating to the new tunnel implementation that will have to be worked through before we do that.

Reviewed By: stepanhruda

Differential Revision: D9566002

fbshipit-source-id: 566feeae426f814206f44bb6284cc1664dcc2774