Add "hideDetails" option to <SettingsControl>
Summary: `<SettingsControls>` by default shows a byline with the long description of a config option. This is useful in the settings view, but can be less useful when showing a single config option in an actual UI. This diff adds an optional param to hide the description in a SettingsControl element. It does this by setting the description to `''`, which seemed like a quick solution.

Reviewed By: tjfryan

Differential Revision: D9615434

fbshipit-source-id: 873fd2f361843e3b8fda7e4542c6fdf1982260d6