Improve 'auto require' heuristics in the case of ties
Summary:
Right now, the auto-require functionality doesn't know what to do when multiple modules are available that match the symbol to be required (e.g. for `FDSCard`, there's `FDSCard.react` and `FDSCard.react.proton`).

In cases of ties, we can employ the additional heuristic of choosing the import with the shortest length. (We should also prioritize .react suffixes over e.g. .rm. I'll put this in `compareForsuggestion` in the next diff.)

Reviewed By: wanderley

Differential Revision: D9621719

fbshipit-source-id: 5510bcf54e43fbfb77e5347cadd51e2b48148efa