Rename fb-OnDemandHostnameUtils.js to hostnames.js
Summary:
I want to use this in other public packages. I don't think it's much of an issue to include this with open source releases.

The other, correct, option is to build a "hostname prettifying service" that each place uses - internally it will do what this function does, the open source release won't have one registered.

Reviewed By: aadsm

Differential Revision: D9683613

fbshipit-source-id: 4aede9499c5ec170b13fba75ca28f0cdde2f98dd