handle symlinks better in TRFS
Summary:
When a remote dir is read, a list of files is obtained. Currently `stat` is called on each file, and stat results are returned to the client. `stat` will follow what a symlink points to. If a symlink points to nothing, the file is not found, which results in an error for the entire operation.

Change this to run `lstat`, which is the same as stat but will return stat info on the symlink itself if a file is a symlink.

Also change logic in RemoteDirectory to render symlinks as files.

Note: the current remote filesystem will *never* return a file of type symlink (see `readdir` in nuclide/pkg/nuclide-fs/lib/FsFileSystem.js). It omits symlinks that point to nothing by try/catching stat. This hid the strange behavior of all symlinks being rendered as directories, which is being changed in this diff.

Reviewed By: semmypurewal

Differential Revision: D9686164

fbshipit-source-id: a3ff05c02df8784fe8c47939a9f969087e5462f2