Fix console bugs
Summary:
If the user manages to send a keystroke while the console is still initializing, the debugger will start at the top of the screen instead of below the shell prompt. This is because we were using the cached cursor position to write the keystroke back out, but we hadn't queried it for the first time yet. Added check to ensure we have a cursor position before using it.

Also, add some logging, and cancel the timeout when we get the cursor position.

Reviewed By: ebluestein

Differential Revision: D9751290

fbshipit-source-id: abd93773b4aa93b897d1f17dd654fa527b1a6e4e