fix issues with queued input
Summary:
calling prompt() on the console shouldn't assume that there isn't any input already. It should just repaint the prompt.

User visible symptom: type ahead before a breakpoint hits. Before the diff, the typeahead input would appear but the insertion point would incorrectly be at the start of the typeahead text. With this fix, the cursor is propertly at the end.

Reviewed By: ebluestein

Differential Revision: D9753309

fbshipit-source-id: dbc735eae5cff7b5ed4d969d8c38b017012e81f8