Truncate the number of diagnostics that can be shown for a given file.
Summary:
NOTE: This adds to an existing type which is part of the public api of Atom IDE, and it changes the behavior associated with a public Atom IDE method. **Any change in behavior will only occur when the number of diagnostics exceeds 1000**.

This adds an additional property to the `DiagnosticsMessages` type indicating the total number of messages associated with the file. It also truncates the number of diagnostics when handing them to observers. Additionally, consumers are informed of the total number of messages, and can determine whether the list is complete with `messages.length === totalMessages`.

**In my tests, this reduces the longest observed stall from the first 15 seconds of typing from 7.4 seconds to around 800ms, a nearly 10x reduction.**

Reviewed By: matthewwithanm

Differential Revision: D9668305

fbshipit-source-id: 9b1a878bd2487c3bea5d8c710d5def024caa3cff