nuclide-commons: add `throttleTime` to `thottle` in observable.js
Summary:
This updates the `throttle` operator in our observables utilities to intelligently call through to RxJS's `throttle` or `throttleTime` operators. It also defaults **both** `leading` and `trailing` options to `true`.

For time-based throttles, it's more useful to use with its options, `leading` and `trailing`. However, they're part
of an options object that's the third argument to the function on the prototype. The second
argument is the scheduler, which is rarely overridden, requiring awkward calls like this:

`observable.throttleTime(300, undefined /* default scheduler */, {leading: true});`

This diff simplifies a call like this into:

`observable.let(throttle(300, {leading: true}));`

... and moves us away from relying on yet another prototype method.

Reviewed By: matthewwithanm

Differential Revision: D9691601

fbshipit-source-id: 4ab08abaa6a55bb8e15479d47b9139574b53cd11