Java Debugger: Fix Bug With Conditional Breakpoints
Summary:
There was a bug where conditional breakpoints could not be changed to a new expression. Additionally, there was a separate bug where conditional expressions where not evaluated because the context manager thought that the debugee was not paused. Both of these have been fixed in this diff.

Note: Expressions like "n < 4" do NOT work but expressions like "4 >= n" do. This seems to be a bug coming from the EvaluationSourceGenerator coming from org.eclipse.jdt.internal.debug.eval.ast.engine. If you inspect the parsableSource variable in the EvaluationManager::evaluateOnCallFrame, we see that the parsableSource variable is WRONG for "n < 4" and not wrong for "4 >= n". For "n < 4" the __run method does NOT return a value whereas for "4 >= n", the __run method DOES return a value. The value we expect. So for "n < 4" we get a void value and also a compilation error which results in the breakpoint ALWAYS hitting even if the condition is not true.

Reviewed By: ebluestein

Differential Revision: D9650008

fbshipit-source-id: 2c36b43cbc17a2fa477606138d217d8a982680fc