Activate when tree-sitter grammars are used
Summary:
This makes the `nuclide-flow` package activate when the flow tree-sitter grammar is used.

There's currently a bug in Atom where the `$PACKAGE_NAME:grammar-used` does *not* activate for tree sitter grammars. Max is fixing this soon.

Add the new, preferred way of activating packages on individual grammar use rather than when any grammar in a package is used, `$GRAMMAR_NAME:root-scope-used`.

Reviewed By: mostafaeweda

Differential Revision: D9761630

fbshipit-source-id: c3bf647020c9ab41e0037550280bfa5963930d0e