Java Debugger: Keep Watching for Class Prepare Events Even If Breakpoint Has Been Set
Summary: This diff makes it so that class prepare requests are not disabled after the breakpoint has been resolved. This means that we will continue getting class prepare events for all breakpoints. This actually fixes some issues with breakpoints not hitting because the exact same class was loaded by separate loaders and the breakpoint request was only set on one of the loaders.

Reviewed By: ebluestein

Differential Revision: D9780129

fbshipit-source-id: 90a49660733eae12caf41e9d446d1d0680ac8dfc